+++ docs/api/groups/constants.h
@@ -1,0 +1,6 @@
+/**
+@defgroup constants Public constants
+
+This group contains all public global constants. Click on the constant name for an explanation.
+
+*/