+++ include/lomse_document_layouter.h
@@ -56,15 +56,18 @@ class DocLayouter : public Layouter
 {
 protected:
     ImoDocument* m_pDoc;
+    LUnits m_pageWidth;
+    LUnits m_pageHeight;
 
     //for unit tests: need to access ScoreLayouter.
     Layouter* m_pScoreLayouter;
 
 public:
-    DocLayouter(Document* pDoc, LibraryScope& libraryScope);
+    DocLayouter(Document* pDoc, LibraryScope& libraryScope, int constrains=0);
     virtual ~DocLayouter();
 
     void layout_document();
+    void layout_empty_document();
 
     //implementation of virtual methods in Layouter base class
     void layout_in_box() {}
@@ -78,6 +81,7 @@ public:
 
 protected:
     void layout_content();
+    void fix_document_size();
 
     GmoBoxDocPage* create_document_page();
     void assign_paper_size_to(GmoBox* pBox);