+++ include/lomse_graphic_view.h
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 // This file is part of the Lomse library.
-// Lomse is copyrighted work (c) 2010-2016. All rights reserved.
+// Lomse is copyrighted work (c) 2010-2018. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -44,8 +44,10 @@
 using namespace std;
 
 
+///@cond INTERNAL
 namespace lomse
 {
+///@endcond
 
 //forward declarations
 class ScreenDrawer;
@@ -78,6 +80,38 @@ typedef SharedPtr<GmoShape>  SpGmoShape;
 //};
 
 
+
+//-----------------------------------------------------------------------------
+/** @ingroup enumerations
+
+    This enum describes the available view types for displaying a document.
+    - @b k_view_simple means that the document will be displayed not paginated,
+        in a single page. It was developed to create small images for
+        controls (i.e. a combobox) by rendering small scores (just one measure)
+        without margins and grey areas (gaps between pages). **[DEPRECATED]**
+    - @b k_view_vertical_book means that the document will be displayed as
+        book pages, one page after the other in a vertical layout. The user will
+        have to scroll down for advancing.
+    - @b k_view_horizontal_book means that the document will be displayed as
+        book pages, one page after the other in a horizontal layout. The user will
+        have to scroll right for advancing.
+    - @b k_view_single_system is for rendering documents that only contain one
+        score (e.g. LDP files, LMD files with just one score, and score files imported
+        from other formats such as MusicXML). It will display the score in a single
+        system, as if the paper had infinite width. And for viewing the end of the score
+        the user will have to scroll to the right. See SingleSystemView.
+
+    @#include <lomse_graphic_view.h>
+*/
+enum EViewType {
+    k_view_simple=0,
+    k_view_vertical_book,
+    k_view_horizontal_book,
+    k_view_single_system,
+};
+
+///@cond INTERNAL
+
 //---------------------------------------------------------------------------------------
 // factory class to create views
 class ViewFactory
@@ -85,8 +119,6 @@ class ViewFactory
 public:
     ViewFactory();
     virtual ~ViewFactory();
-
-    enum { k_view_simple=0, k_view_vertical_book, k_view_horizontal_book, };
 
     static View* create_view(LibraryScope& libraryScope, int viewType,
                              ScreenDrawer* pDrawer);
@@ -107,8 +139,12 @@ struct PageRectangle
     }
 };
 
+///@endcond
+
 //---------------------------------------------------------------------------------------
-// A view to edit the score in full page
+/** %GraphicView is an abstract base class for Views rendering the document as a
+    graphic.
+*/
 class GraphicView : public View
 {
 protected:
@@ -151,10 +187,18 @@ protected:
     RenderingBuffer* m_pPrintBuf;
     double           m_print_ppi;     //printer resolution in pixels per inch
 
+    //options
+    Color       m_backgroundColor;
+
 public:
+///@cond INTERNALS
+//excluded from public API because the View methods are managed from Interactor
+
     virtual ~GraphicView();
 
-    //view settings
+    /// @name View settings
+    ///@{
+
     void new_viewport(Pixels x, Pixels y);
     void set_rendering_buffer(RenderingBuffer* rbuf);
     void get_viewport(Pixels* x, Pixels* y) { *x = m_vxOrg; *y = m_vyOrg; }
@@ -165,7 +209,11 @@ public:
     void add_visual_effect(VisualEffect* pEffect);
     void set_visual_effects_for_mode(int mode);
 
-    //renderization related
+    ///@}    //View settings
+
+
+    /// @name Renderization related
+    ///@{
     VRect get_damaged_rectangle();
     UPoint get_page_origin_for(GmoObj* pGmo);
     void draw_all_visual_effects();
@@ -175,32 +223,58 @@ public:
     void draw_dragged_image();
     void draw_selected_objects();
     void draw_handler(Handler* pHandler);
+    void set_background(Color color) { m_backgroundColor = color; }
 
-    //scrolling support
+    ///@}    //Renderization related
+
+
+    /// @name Scrolling support
+    ///@{
     virtual void get_view_size(Pixels* xWidth, Pixels* yHeight) = 0;
     virtual void change_viewport_if_necessary(ImoId id);
 
-    //selection rectangle
+    ///@}    //Scrolling support
+
+
+    /// @name Selection rectangle
+    ///@{
     void start_selection_rectangle(LUnits x1, LUnits y1);
     void hide_selection_rectangle();
     void update_selection_rectangle(LUnits x2, LUnits y2);
 
-    //tempo line
+    ///@}    //Selection rectangle
+
+
+    /// @name Tempo line
+    ///@{
     void show_tempo_line(Pixels x1, Pixels y1, Pixels x2, Pixels y2);
     void hide_tempo_line();
     void update_tempo_line(Pixels x2, Pixels y2);
 
-    //highlighting notes and rests
+    ///@}    //Tempo line
+
+
+    /// @name Highlighting notes and rests
+    ///@{
     void highlight_object(ImoStaffObj* pSO);
     void remove_highlight_from_object(ImoStaffObj* pSO);
     void remove_all_highlight();
 
-    // The View is requested to re-paint itself onto the window
+    ///@}    //Highlighting notes and rests
+
+
+    /** The View is requested to re-paint itself onto the window */
     virtual void redraw_bitmap();
 
-    //inline DocCursor& get_cursor() { return m_cursor; }
+    //graphical model
+    GraphicModel* get_graphic_model();
 
-    //caret
+    //handlers
+    Handler* handlers_hit_test(LUnits x, LUnits y);
+
+
+    /// @name Caret
+    ///@{
     void show_caret();
     void hide_caret();
     void toggle_caret();
@@ -211,19 +285,21 @@ public:
     bool is_caret_blink_enabled();
     void change_cursor_voice(int voice);
 
-    //dragged image associated to mouse cursor
+    ///@}    //Caret
+
+
+    /// @name Dragged image associated to mouse cursor
+    ///@{
     void move_drag_image(LUnits x, LUnits y);
     void set_drag_image(GmoShape* pShape, bool fGetOwnership, UPoint offset);
     void show_drag_image(bool value);
     void enable_drag_image(bool fEnabled);
 
-    //handlers
-    Handler* handlers_hit_test(LUnits x, LUnits y);
+    //@}    //Dragged image associated to mouse cursor
 
-    //graphical model
-    GraphicModel* get_graphic_model();
 
-    //coordinates conversions
+    /// @name Coordinates conversion
+    ///@{
     void screen_point_to_page_point(double* x, double* y);
     void model_point_to_screen(double* x, double* y, int iPage);
     UPoint screen_point_to_model_point(Pixels x, Pixels y);
@@ -235,7 +311,11 @@ public:
                                                      list<PageRectangle*>* rectangles);
     LUnits pixels_to_lunits(Pixels pixels);
 
-    //scale
+    ///@}    //Coordinates conversion
+
+
+    /// @name Scale
+    ///@{
     void zoom_in(Pixels x=0, Pixels y=0);
     void zoom_out(Pixels x=0, Pixels y=0);
     void zoom_fit_full(Pixels width, Pixels height);
@@ -244,19 +324,43 @@ public:
     double get_scale();
     double get_resolution();
 
-    //rendering options
+    ///@}    //Scale
+
+
+    /// @name Rendering options
+    ///@{
     void set_rendering_option(int option, bool value);
     void reset_boxes_to_draw();
     void set_box_to_draw(int boxType);
     void highlight_voice(int voice);
 
-    //support for printing
+    ///@}    //Rendering options
+
+
+    /// @name Layout constrains
+    ///@{
+    virtual int get_layout_constrains() = 0;
+    virtual bool is_valid_for_this_view(Document* pDoc) = 0;
+
+    ///@}    //Layout constrains
+
+
+    /// @name Support for printing
+    ///@{
     void set_print_buffer(RenderingBuffer* rbuf) { m_pPrintBuf = rbuf; }
     void set_print_ppi(double ppi) { m_print_ppi = ppi; }
     virtual void print_page(int page, VPoint viewport);
 
+    ///@}    //Support for printing
+
+
     //info
     AreaInfo* get_info_for_point(Pixels x, Pixels y);
+
+
+
+
+///@endcond
 
 protected:
     GraphicView(LibraryScope& libraryScope, ScreenDrawer* pDrawer);
@@ -292,34 +396,51 @@ protected:
 };
 
 
+///@cond INTERNALS
+
 //---------------------------------------------------------------------------------------
-// A graphic view with one page, no margins (i.e. LenMus ScoreAuxCtrol)
+/** %SimpleView is a GraphicView for displaying the document in a single page, without
+    borders and margins.
+*/
 class LOMSE_EXPORT SimpleView : public GraphicView
 {
 public:
+
     SimpleView(LibraryScope& libraryScope, ScreenDrawer* pDrawer);
     virtual ~SimpleView() {}
 
     virtual int page_at_screen_point(double x, double y);
     void set_viewport_for_page_fit_full(Pixels screenWidth);
     void get_view_size(Pixels* xWidth, Pixels* yHeight);
+    virtual int get_layout_constrains() { return k_use_paper_width | k_use_paper_height; }
+    bool is_valid_for_this_view(Document* UNUSED(pDoc)) { return true; }
 
 protected:
     void collect_page_bounds();
 
 };
+///@endcond
 
 
 //---------------------------------------------------------------------------------------
-// A graphic view with pages in vertical (i.e. Adobe PDF Reader, MS Word)
+/** %VerticalBookView is a GraphicView for rendering documents in pages, with the
+    pages spread in vertical (i.e. Adobe PDF Reader, MS Word)
+*/
 class LOMSE_EXPORT VerticalBookView : public GraphicView
 {
 public:
+///@cond INTERNALS
+//excluded from public API because the View methods are managed from Interactor
+
     VerticalBookView(LibraryScope& libraryScope, ScreenDrawer* pDrawer);
     virtual ~VerticalBookView() {}
 
     void set_viewport_for_page_fit_full(Pixels screenWidth);
     void get_view_size(Pixels* xWidth, Pixels* yHeight);
+    virtual int get_layout_constrains() { return k_use_paper_width | k_use_paper_height; }
+    bool is_valid_for_this_view(Document* UNUSED(pDoc)) { return true; }
+
+///@endcond
 
 protected:
     void collect_page_bounds();
@@ -328,17 +449,96 @@ protected:
 
 
 //---------------------------------------------------------------------------------------
-// A graphic view with pages in horizontall (i.e. Finale, Sibelius)
+/** %HorizontalBookView is a GraphicView for rendering documents in pages, with the
+    pages spread in horizontal (i.e. Finale, Sibelius)
+*/
 class LOMSE_EXPORT HorizontalBookView : public GraphicView
 {
 protected:
 
 public:
+///@cond INTERNALS
+//excluded from public API because the View methods are managed from Interactor
+
     HorizontalBookView(LibraryScope& libraryScope, ScreenDrawer* pDrawer);
     virtual ~HorizontalBookView() {}
 
     void set_viewport_for_page_fit_full(Pixels screenWidth);
     void get_view_size(Pixels* xWidth, Pixels* yHeight);
+    virtual int get_layout_constrains() { return k_use_paper_width | k_use_paper_height; }
+    bool is_valid_for_this_view(Document* UNUSED(pDoc)) { return true; }
+
+///@endcond
+
+protected:
+    void collect_page_bounds();
+
+};
+
+
+//---------------------------------------------------------------------------------------
+/** %SingleSystemView is a GraphicView for rendering documents that only contain one
+    score (e.g. LDP files, LMD files with just one score, and score files imported
+    from other formats such as MusicXML). If the document does not contains scores
+    or contains more than one score, this view will display an empty view.
+
+    This view will display the score in a single system, as if the paper had infinite
+    width. And for viewing the end of the score the user will have to scroll to the
+    right.
+
+    When the displayed score does not end in barline but the staff lines continue
+    running until the end of the page, the staff lines will be finished after running
+    empty for the length of last occupied measure.
+
+
+    <b>Margins</b>
+
+    The score is displayed on a white paper and the margins around the score are
+    as follows:
+    - Top margin = document top margin + score top margin
+    - Left margin = document left margin + score left margin
+    - Bottom margin = Top margin
+    - Right margin = Left margin
+
+
+    <b>Background color</b>
+
+    The white paper is surrounded by the background (gray color). As with all Views,
+    the background color can be changed by invoking Interactor::set_view_background().
+    E.g. for suppressing the background:
+    @code
+        m_pPresenter = lomse.open_document(k_view_single_system, filename);
+        if (SpInteractor spInteractor = m_pPresenter->get_interactor(0).lock())
+        {
+            spInteractor->set_rendering_buffer(&m_rbuf_window);
+            spInteractor->set_view_background( Color(255,255,255) );  //white
+            ...
+    @endcode
+
+
+    <b>AutoScroll</b>
+
+    As playback advances the View generates EventUpdateViewport events so that measure
+    being played is always totally visible.
+
+*/
+class LOMSE_EXPORT SingleSystemView : public GraphicView
+{
+public:
+///@cond INTERNALS
+//excluded from public API because the View methods are managed from Interactor
+
+    SingleSystemView(LibraryScope& libraryScope, ScreenDrawer* pDrawer);
+    virtual ~SingleSystemView() {}
+
+    virtual int page_at_screen_point(double x, double y);
+
+    void set_viewport_for_page_fit_full(Pixels screenWidth);
+    void get_view_size(Pixels* xWidth, Pixels* yHeight);
+    virtual int get_layout_constrains() { return k_infinite_width | k_use_paper_height; }
+    bool is_valid_for_this_view(Document* pDoc);
+
+///@endcond
 
 protected:
     void collect_page_bounds();