+++ src/graphic_model/layouters/lomse_score_layouter.cpp
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 // This file is part of the Lomse library.
-// Lomse is copyrighted work (c) 2010-2017. All rights reserved.
+// Lomse is copyrighted work (c) 2010-2018. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -285,6 +285,7 @@ void ScoreLayouter::create_system_layouter()
     m_pCurSysLyt = LOMSE_NEW SystemLayouter(this, m_libraryScope, m_pScoreMeter,
                                       m_pScore, m_shapesStorage, m_pShapesCreator,
                                       m_pPartsEngraver, m_pSpAlgorithm);
+    m_pCurSysLyt->set_constrains(m_constrains);
     m_sysLayouters.push_back(m_pCurSysLyt);
 }
 
@@ -530,15 +531,6 @@ LUnits ScoreLayouter::distance_to_top_of_system(int iSystem, bool fFirstInPage)
 }
 
 //---------------------------------------------------------------------------------------
-void ScoreLayouter::determine_staff_lines_horizontal_position(int iInstr)
-{
-    LUnits indent = get_system_indent();
-    LUnits width = m_pCurBoxSystem->get_usable_width();
-    m_pPartsEngraver->set_staves_horizontal_position(iInstr, m_cursor.x, width, indent);
-    m_cursor.x += indent;
-}
-
-//---------------------------------------------------------------------------------------
 LUnits ScoreLayouter::determine_top_space(int iInstr, bool fFirstSystemInScore,
                                           bool fFirstSystemInPage)
 {
@@ -686,7 +678,9 @@ void ScoreLayouter::fill_page_with_empty_systems_if_required()
     if (!m_pCurSysLyt->system_must_be_truncated())
     {
         ImoOptionInfo* pOpt = m_pScore->get_option("Score.FillPageWithEmptyStaves");
-        bool fFillPage = pOpt->get_bool_value();
+       bool fFillPage = pOpt->get_bool_value()
+                         && !((m_constrains & k_infinite_height)
+                              || (m_constrains & k_infinite_width));
         if (fFillPage)
         {
             while(enough_space_for_empty_system())